backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
sqlalchemy>=2.0.36
python-docx>=1.1.2
jinja2>=3.1.4
python-dotenv>=1.0.1
aiosqlite>=0.20.0

